doc_pat/requirements.txt
asgiref==3.2.10
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
django-import-export==2.6.1
django-phonenumber-field==5.0.0
djangorestframework==3.12.1
et-xmlfile==1.1.0
gunicorn==20.0.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
phonenumbers==8.12.11
pytz==2020.1
PyYAML==6.0
six==1.15.0
sqlparse==0.4.1
tablib==3.1.0
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
django-import-export
django-phonenumber-field
# weasyprint
xhtml2pdf